getter for enemy was returning double and not int so i fixed that;

+++ src/main/java/board/Enemy.java
@@ -43,11 +43,11 @@ public class Enemy {
 	}
 	
 	
-	public double getEnemyX() {
+	public int getEnemyX() {
 		return enemyX;
 	}
 
-	public double getEnemyY() {
+	public int getEnemyY() {
 		return enemyY;
 	}
